10/22 NameGenerator work
It's now (TESTED), with the basic functionality of the class working.
All that's left to do is add in the different name selections for each
race, and testing those out (though I see no issue with it).

From there are some minor changes down the line, about how specifically
the race will be chosen. More to think about later..