A couple of new files
CharacterRace, for determining the character's race randomly. Gives the default stats/bonuses for each race in a dict that should be similar to another one down the line. 

__init__, after a small amount of researching, is included to allow importing from the same directory, and should (hopefully) make things easier moving forward. 

For CharacterRace, the default stats need to be added in in order for the 'stats' dict to be updated properly, much like how the names still need to be added in.